textclust/dashboard/requirements.txt
astroid==2.5
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dash[celery]==2.3.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-bootstrap-components==0.12.0
dash-renderer==1.9.0
elasticsearch==7.11.0
Flask==1.1.2
Flask-Compress==1.9.0
Flask-Cors==3.0.10
future==0.18.2
gensim==3.8.3
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonpickle==2.0.0
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.5
numpy==1.22.2
pandas==1.2.2
plotly==5.3.1
pylint==2.7.1
pymongo==3.11.3
pyorbital==1.6.0
PySocks==1.7.1
python-dateutil==2.8.1
regex==2020.11.13
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.57.0
tweepy==4.8.0
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
python-dotenv==0.18.0
gunicorn==20.1.0
spacy==3.2.3
dtaidistance==2.3.6
emoji==1.6.3
celery==5.2.3
redis==4.2.0
networkx==2.6.3
python-louvain==0.16
matplotlib==3.5.1
pyarrow==7.0.0
